remove apply pages
